empresa: funcoes buscam por id ou codemp

+++ database/crud/empresa.py
@@ -1,7 +1,7 @@
 from database.database import AsyncSessionLocal
 from database.models import Empresa
 from src.utils.log import Log
-from src.utils.db import validar_dados
+from src.utils.db import validar_dados, formatar_retorno
 from sqlalchemy.future import select
 import os
 import logging
@@ -57,7 +57,14 @@ async def criar(snk_codemp:int,
         await session.refresh(nova_empresa)
         return True
 
-async def atualizar_id(empresa_id:int,**kwargs):
+async def atualizar(
+        id:int=None,
+        codemp:int=None,
+        **kwargs
+    ):
+
+    if not any([id,codemp]):
+        return False
 
     if kwargs:
         kwargs = validar_dados(modelo=Empresa,
@@ -67,9 +74,15 @@ async def atualizar_id(empresa_id:int,**kwargs):
             return False
 
     async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(Empresa).where(Empresa.id == empresa_id)
-        )
+        if id:
+            result = await session.execute(
+                select(Empresa).where(Empresa.id == id)
+            )
+
+        if codemp:
+            result = await session.execute(
+                select(Empresa).where(Empresa.snk_codemp == codemp)
+            )
         empresa = result.scalar_one_or_none()
         if not empresa:
             return False
@@ -79,70 +92,60 @@ async def atualizar_id(empresa_id:int,**kwargs):
         await session.refresh(empresa)
         return True
 
-async def atualizar_codigo(snk_codemp:int,**kwargs):
+async def buscar(
+        id:int=None,
+        codemp:int=None
+    ):
 
-    if kwargs:
-        kwargs = validar_dados(modelo=Empresa,
-                            kwargs=kwargs,
-                            colunas_criptografadas=COLUNAS_CRIPTOGRAFADAS)
-        if not kwargs:
-            return False
+    if not any([id,codemp]):
+        return False    
 
     async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(Empresa).where(Empresa.snk_codemp == snk_codemp)
-        )
+
+        if id:
+            result = await session.execute(
+                select(Empresa).where(Empresa.id == id)
+            )
+
+        if codemp:
+            result = await session.execute(
+                select(Empresa).where(Empresa.snk_codemp == codemp)
+            )
+
         empresa = result.scalar_one_or_none()
         if not empresa:
+            print(f"Empresa não encontrada. Parâmetro: {codemp or id}")
             return False
-        for key, value in kwargs.items():
-            setattr(empresa, key, value)
-        await session.commit()
-        await session.refresh(empresa)
-        return True
+        
+        dados_empresa = formatar_retorno(colunas_criptografadas=COLUNAS_CRIPTOGRAFADAS,
+                                         retorno=empresa)
+        
+        return dados_empresa
 
-async def buscar_codigo(snk_codemp:int):
+async def excluir(
+        id:int=None,
+        codemp:int=None
+    ):
+
+    if not any([id,codemp]):
+        return False    
 
     async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(Empresa).where(Empresa.snk_codemp == snk_codemp)
-        )
-        empresa = result.scalar_one_or_none()
-        if not empresa:
-            print(f"Empresa não encontrada. Parâmetro: {snk_codemp}")
-            return False
-        return empresa.__dict__
+        if id:
+            result = await session.execute(
+                select(Empresa).where(Empresa.id == id)
+            )
 
-async def excluir_id(id:int):
+        if codemp:
+            result = await session.execute(
+                select(Empresa).where(Empresa.snk_codemp == codemp)
+            )
 
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(Empresa).where(Empresa.id == id)
-        )
         empresa = result.scalar_one_or_none()
         if not empresa:
             print(f"Empresa não encontrada. Parâmetro: {id}")
             return False
-        try:
-            await session.delete(empresa)
-            await session.commit()
-            return True
-        except Exception as e:
-            await session.rollback()
-            print("Erro ao excluir empresa no banco de dados: %s", e)
-            logger.error("Erro ao excluir empresa no banco de dados: %s", e)
-            return False
-
-async def excluir_codigo(snk_codemp:int):
-
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(Empresa).where(Empresa.snk_codemp == snk_codemp)
-        )
-        empresa = result.scalar_one_or_none()
-        if not empresa:
-            print(f"Empresa não encontrada. Parâmetro: {snk_codemp}")
-            return False
+        
         try:
             await session.delete(empresa)
             await session.commit()
